Adjust for single precision.

+++ Source/FieldSolver/FiniteDifferenceSolver/EvolveB.cpp
@@ -198,7 +198,7 @@ void FiniteDifferenceSolver::EvolveBCartesian (
             amrex::Gpu::synchronize();
             wt = amrex::second() - wt;
             amrex::HostDevice::Atomic::Add( &(*cost)[mfi.index()], wt);
-            amrex::HostDevice::Atomic::Add( &(*(warpx.g_temp)[lev])[mfi.index()], wt);
+            amrex::HostDevice::Atomic::Add( &(*(warpx.g_temp)[lev])[mfi.index()], double(wt));
         }
     }
